リファクタリング: DocRowコンポーネントのロジックを簡素化し、カスタムフックを使用して編集モードとフェードアウト管理を改善
